+++ R/print.R
@@ -1,97 +1,65 @@
 #' Print an incidence object.
 #'
 #' @param x An 'incidence' object.
-#' @param ... Not used.
+#' @param ... Additional arguments passed through to the tibble format method.
 #'
+#' @name print_incidence
 #' @export
 print.incidence <- function(x, ...) {
-
-  # get the date and count variables
-  count_var <- get_count_names(x)
-  date_var <- get_dates_name(x)
-  dat <- x[[date_var]]
-
-  # header
-  header <- sprintf("An incidence object: %s x %s\n",
-                    formatC(nrow(x), big.mark = ","),
-                    formatC(ncol(x), big.mark = ","))
-  cat(pillar::style_subtle(header))
-
-  # print tibble but remove title
-  out <- format(tibble::as_tibble(x))
-  cat(out[-1], sep = "\n")
-  cat("\n")
-
+  writeLines(format(x, ...))
   invisible(x)
 }
 
 
-#' Print an incidence object.
-#'
-#' @param x An 'incidence2' object.
-#' @param ... Not used.
-#'
 #' @export
-print.incidence2 <- function(x, ...) {
+#' @rdname print_incidence
+format.incidence <- function(x, ...) {
+  overview <- overview(x)
+  body <- format(tibble::as_tibble(x, ...))[-1]
+  if (inherits(x, "incidence2")) {
+    inter <- interval(x)
+    out <- c(overview, inter, "", body)
+  } else {
+    out <- c(overview, "", body)
+  }
+  out
+}
 
-  # get the date and count variables
+
+overview <- function(x) {
   count_var <- get_count_names(x)
   date_var <- get_dates_name(x)
   dat <- x[[date_var]]
+  d1 <- min(dat)
+  d2 <- max(dat)
+  header <- sprintf(
+    "An incidence object: %s x %s",
+    formatC(nrow(x), big.mark = ","),
+    formatC(ncol(x), big.mark = ",")
+  )
+  header <- pillar::style_subtle(header)
+  date_range <- sprintf("date range: %s to %s", d1, d2)
+  cases <- vapply(
+    count_var,
+    function(var) {
+      if(var == "count") {
+        sprintf("cases: %d", sum(x[[var]]))
+      } else {
+        sprintf("%s: %d", var, sum(x[[var]]))
+      }
+    },
+    character(1)
+  )
+  c(header, date_range, cases)
+}
 
-  # header
-  header <- sprintf("An incidence object: %s x %s\n",
-                    formatC(nrow(x), big.mark = ","),
-                    formatC(ncol(x), big.mark = ","))
-  cat(pillar::style_subtle(header))
-
-  # cases over date range
-  for (i in count_var) {
-    if (inherits(dat, "Date") ||
-        inherits(dat, "numeric") ||
-        inherits(dat, "grates_yearweek") ||
-        inherits(dat, "grates_quarter") ||
-        inherits(dat, "grates_year") ||
-        (inherits(dat, "grates_month") && attr(dat, "n") == 1)) {
-      d1 <- min(dat)
-      d2 <- max(dat)
-    } else if (inherits(dat, "grates_int_period") || inherits(dat, "numeric") || inherits(dat, "integer")) {
-      d1 <- as.integer(min(dat))
-      d2 <- as.integer(max(dat + 1)) - 1
-    } else {
-      d1 <- as.Date(min(dat))
-      d2 <- as.Date(max(dat + 1)) - 1
-    }
-
-    if(i == "count") {
-      msg <- sprintf("%d cases from %s to %s\n", sum(x[[i]]), d1, d2)
-    } else {
-      msg <- sprintf("%d %s from %s to %s\n", sum(x[[i]]), i, d1, d2)
-    }
-    cat(msg)
-  }
-
-
-  # interval
+interval <- function(x) {
   interval <- get_interval(x)
   if (is.numeric(interval)) {
-    cat(sprintf("interval: %d %s\n", interval, ifelse(interval < 2, "day", "days")))
+    sprintf("interval: %d %s", interval, ifelse(interval < 2, "day", "days"))
   } else if (grepl("\\d", interval)) {
-    cat(sprintf("interval: %s\n", interval))
+    sprintf("interval: %s", interval)
   } else {
-    cat(sprintf("interval: 1 %s\n", interval))
+    sprintf("interval: 1 %s", interval)
   }
-
-  # cumulative
-  cumulative <- attr(x, "cumulative")
-  if (!is.null(cumulative)) {
-    cat(sprintf("cumulative: %s\n\n", cumulative))
-  }
-
-  # print tibble but remove title
-  out <- format(tibble::as_tibble(x))
-  cat(out[-1], sep = "\n")
-  cat("\n")
-
-  invisible(x)
 }